nvagetbios: Support cards with BIOS higher than first 4GB of physical VRAM.

+++ nvagetbios.c
@@ -155,10 +155,10 @@ int vbios_extract_pramin(int cnum, uint8_t *vbios, int *length)
 			cnum, nva_cards[cnum].chipset);
 
 	if (nva_cards[cnum].card_type >= 0x50) {
-		uint32_t vbios_vram = (nva_rd32(cnum, 0x619f04) & ~0xff) << 8;
+		uint64_t vbios_vram = (uint64_t)(nva_rd32(cnum, 0x619f04) & ~0xff) << 8;
 
 		if (!vbios_vram)
-			vbios_vram = (nva_rd32(cnum, 0x1700) << 16) + 0xf0000;
+			vbios_vram = ((uint64_t)nva_rd32(cnum, 0x1700) << 16) + 0xf0000;
 
 		old_bar0_pramin = nva_rd32(cnum, 0x1700);
 		nva_wr32(cnum, 0x1700, vbios_vram >> 16);
